app/requirements.txt
backports.csv==1.0.7
celery==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
django-extensions==2.1.9
dj-database-url==0.5.0
django-environ==0.4.5
django-filter==2.2.0
Django==3.0.8
django-import-export==1.2.0
django-celery-email==2.0.2
django-tables2==2.0.6
django-bootstrap3==11.1.0
django-bootstrap4==0.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-modal-forms==2.0.0
django-widget-tweaks==1.4.2
django-heroku==0.3.1
et-xmlfile==1.0.1
flower==0.9.5
gunicorn==20.0.4
honcho==0.5.0
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
pytz==2019.1
PyYAML==5.1.1
psycopg2==2.8.3
pandas==0.23.4
pinyin==0.4.0
redis==3.2.1
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
whitenoise==4.1.2
